+++ flowless-sample-single-root/src/main/java/com/zhuinden/flowless_dispatcher_sample/FirstView.java
@@ -80,7 +80,7 @@ public class FirstView
     }
 
     @Override
-    public void onViewRestored(boolean forcedWithBundler) {
+    public void onViewRestored() {
         Log.i(TAG, "onViewRestored()");
     }