+++ flowless-viewpager-example/src/main/java/com/zhuinden/flowless_viewpager_example/FirstView.java
@@ -71,7 +71,7 @@ public class FirstView extends LinearLayout implements FlowLifecycles.ViewLifecy
     }
 
     @Override
-    public void onViewRestored(boolean forcedWithBundler) {
+    public void onViewRestored() {
         Log.i(TAG, "View restored in [" + TAG + "]");
     }